Refactor: recoil -> states
