+++ package.json
@@ -91,8 +91,8 @@
       "^@constants/(.*)$": "<rootDir>/src/constants/$1",
       "^@libs$": "<rootDir>/src/libs",
       "^@libs/(.*)$": "<rootDir>/src/libs/$1",
-      "^@recoil$": "<rootDir>/src/recoil",
-      "^@recoil/(.*)$": "<rootDir>/src/recoil/$1"
+      "^@states$": "<rootDir>/src/states",
+      "^@states/(.*)$": "<rootDir>/src/states/$1"
     },
     "setupFilesAfterEnv": [
       "<rootDir>/jest.setup.js"
