+++ src/components/InfoWindow/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/button-has-type */
 import { useEffect } from 'react';
 import styled from '@emotion/styled';
-import { tabStateAtom } from '@recoil/infoWindowState';
+import { tabStateAtom } from '@states/infoWindowState';
 import { useRecoilState } from 'recoil';
 import { TabState } from '@libs/types/infowindow';
 import { CancelIcon, ClockIcon, CopyIcon, MarkerIcon, PhoneIcon } from '@assets';
