+++ src/components/MapComp/index.tsx
@@ -4,8 +4,8 @@ import { css } from '@emotion/react';
 import PointMarker from '@components/Marker/pointMarker';
 import { CustomGeoJSONFeatures } from '@libs/types/map';
 import { renderEmotionElementToHtml } from '@libs/utils/renderEmotionElementToHtml';
-import { geoJsonSelector } from '@recoil/map';
-import { activeFilterIdAtom } from '@recoil/ui';
+import { geoJsonSelector } from '@states/map';
+import { activeFilterIdAtom } from '@states/ui';
 import mapboxgl, { GeoJSONSourceRaw, Marker } from 'mapbox-gl';
 import { FeatureCollection } from 'geojson';
 import 'mapbox-gl/dist/mapbox-gl.css';
