+++ src/index.tsx
@@ -1,9 +1,9 @@
-import ReactDOM from "react-dom/client";
-import { RecoilRoot } from "recoil";
-import "./index.css";
-import App from "./App";
+import ReactDOM from 'react-dom/client';
+import { RecoilRoot } from 'recoil';
+import './index.css';
+import App from './App';
 
-const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <RecoilRoot>
     <App />
